code cleanups via IntelliJ IDEA inspections

+++ h2/src/main/org/h2/value/ValueFloat.java
@@ -139,7 +139,7 @@ public class ValueFloat extends Value {
 
     @Override
     public Object getObject() {
-        return Float.valueOf(value);
+        return value;
     }
 
     @Override
